migrated utils to disaggregation_fixture.py
