+++ python/sglang/test/server_fixtures/disaggregation_fixture.py
@@ -7,6 +7,8 @@ import warnings
 from typing import ClassVar, Optional
 from urllib.parse import urlparse
 
+import requests
+
 from sglang.srt.environ import envs
 from sglang.srt.utils import kill_process_tree
 from sglang.test.test_utils import (
@@ -21,6 +23,27 @@ from sglang.test.test_utils import (
 from sglang.utils import wait_for_http_ready
 
 logger = logging.getLogger(__name__)
+
+
+def configure_nixl_pd_backend(test_cls):
+    test_cls.transfer_backend = ["--disaggregation-transfer-backend", "nixl"]
+    # NIXL backend/network selection is driven by NIXL environment variables
+    # such as SGLANG_DISAGGREGATION_NIXL_BACKEND and backend params, not by the
+    # Mooncake-specific --disaggregation-ib-device argument.
+    test_cls.rdma_devices = []
+
+
+def assert_process_healthy(test_case, name, process, url, health_path="/health"):
+    test_case.assertIsNotNone(process, f"{name} process was not started")
+    test_case.assertIsNone(
+        process.poll(),
+        f"{name} exited unexpectedly with code {process.returncode}",
+    )
+    try:
+        response = requests.get(f"{url}{health_path}", timeout=10)
+    except requests.RequestException as e:
+        test_case.fail(f"Failed to connect to {name} health endpoint: {e}")
+    test_case.assertEqual(response.status_code, 200, response.text)
 
 
 class PDDisaggregationServerBase(CustomTestCase):
